feat(frontend): component moved, added crosschain check
