+++ packages/frontend/components/Borrow/Header.tsx
@@ -2,7 +2,7 @@ import { Divider, Typography, Stack, Box } from "@mui/material"
 import { ActionType } from "../../helpers/assets"
 import { NetworkIcon } from "../Shared/Icons"
 import TabChip from "../Shared/TabChip"
-import SlippageSettings from "./SlippageSettings"
+import SlippageSettings from "../Shared/SlippageSettings"
 import TooltipWrapper from "../Shared/Tooltips/TooltipWrapper"
 
 type BorrowHeaderProps = {
@@ -10,6 +10,7 @@ type BorrowHeaderProps = {
   actionType: ActionType
   chainName: string
   onActionTypeChange: (action: ActionType) => void
+  isCrossChainOperation: boolean
 }
 
 function BorrowHeader({
@@ -17,6 +18,7 @@ function BorrowHeader({
   actionType,
   chainName,
   onActionTypeChange,
+  isCrossChainOperation,
 }: BorrowHeaderProps) {
   const networkMessage = `Your position is currently on the ${chainName} Network`
 
@@ -32,23 +34,35 @@ function BorrowHeader({
           <Typography variant="body2" height="40px" lineHeight="40px">
             Manage your position
           </Typography>
-          <TooltipWrapper
-            defaultOpen
-            placement="top"
-            title={
-              <Box
-                sx={{
-                  display: "flex",
-                  flexDirection: "column",
-                  width: "10rem",
-                }}
-              >
-                <Typography variant="small">{networkMessage}</Typography>
-              </Box>
-            }
+          <Stack
+            direction="row"
+            justifyContent="space-between"
+            alignItems="center"
+            height="40px"
           >
-            <NetworkIcon network={chainName} height={18} width={18} />
-          </TooltipWrapper>
+            <TooltipWrapper
+              defaultOpen
+              placement="top"
+              title={
+                <Box
+                  sx={{
+                    display: "flex",
+                    flexDirection: "column",
+                    width: "10rem",
+                  }}
+                >
+                  <Typography variant="small">{networkMessage}</Typography>
+                </Box>
+              }
+            >
+              <NetworkIcon network={chainName} height={18} width={18} />
+            </TooltipWrapper>
+            {isCrossChainOperation && (
+              <Box sx={{ ml: "1rem" }}>
+                <SlippageSettings />
+              </Box>
+            )}
+          </Stack>
         </Stack>
       ) : (
         <Stack
@@ -60,7 +74,7 @@ function BorrowHeader({
           <Typography variant="body2" height="40px" lineHeight="40px">
             Borrow
           </Typography>
-          <SlippageSettings />
+          {isCrossChainOperation && <SlippageSettings />}
         </Stack>
       )}
       <Divider sx={{ mt: "1rem", mb: "0.5rem" }} />
